# openapi_output.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smith 2.0",
        "description": "Agente de Vendas Inteligente com IA",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/projects/": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create Project",
                "description": "Cria um novo projeto\n\nArgs:\n    project_data: Dados do projeto a criar\n\nReturns:\n    Projeto criado com ID",
                "operationId": "create_project_api_projects__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__models__project__ProjectCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List Projects",
                "description": "Lista projetos com filtros opcionais\n\nArgs:\n    status: Filtrar por status\n    prioridade: Filtrar por prioridade\n    cliente_id: Filtrar por cliente\n    responsavel: Filtrar por respons\u00c3\u00a1vel\n    limit: N\u00c3\u00bamero m\u00c3\u00a1ximo de resultados\n    offset: Offset para pagina\u00c3\u00a7\u00c3\u00a3o\n\nReturns:\n    Lista de projetos",
                "operationId": "list_projects_api_projects__get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/app__models__project__ProjectStatus"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "prioridade",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/ProjectPriority"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Prioridade"
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Cliente Id"
                        }
                    },
                    {
                        "name": "responsavel",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Responsavel"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/app__models__project__Project"
                                    },
                                    "title": "Response List Projects Api Projects  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/projects/{project_id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get Project",
                "description": "Busca um projeto por ID\n\nArgs:\n    project_id: ID do projeto\n\nReturns:\n    Projeto encontrado",
                "operationId": "get_project_api_projects__project_id__get",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/app__models__project__Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update Project",
                "description": "Atualiza um projeto existente\n\nArgs:\n    project_id: ID do projeto\n    project_data: Dados a atualizar\n\nReturns:\n    Projeto atualizado",
                "operationId": "update_project_api_projects__project_id__put",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__models__project__ProjectUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete Project",
                "description": "Deleta um projeto\n\nArgs:\n    project_id: ID do projeto\n\nReturns:\n    Confirma\u00c3\u00a7\u00c3\u00a3o de dele\u00c3\u00a7\u00c3\u00a3o",
                "operationId": "delete_project_api_projects__project_id__delete",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/portal/projects": {
            "get": {
                "tags": [
                    "Portal do Cliente",
                    "Portal do Cliente"
                ],
                "summary": "List My Projects",
                "description": "Listar projetos do cliente logado",
                "operationId": "list_my_projects_api_portal_projects_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/app__models__client_portal__Project"
                                    },
                                    "title": "Response List My Projects Api Portal Projects Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/portal/projects/{project_id}": {
            "get": {
                "tags": [
                    "Portal do Cliente",
                    "Portal do Cliente"
                ],
                "summary": "Get Project",
                "description": "Buscar projeto com detalhes",
                "operationId": "get_project_api_portal_projects__project_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectWithDetails"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/portal/admin/projects": {
            "get": {
                "tags": [
                    "Portal do Cliente",
                    "Portal do Cliente"
                ],
                "summary": "List All Projects",
                "description": "Listar todos os projetos (admin)",
                "operationId": "list_all_projects_api_portal_admin_projects_get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/app__models__client_portal__Project"
                                    },
                                    "title": "Response List All Projects Api Portal Admin Projects Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Portal do Cliente",
                    "Portal do Cliente"
                ],
                "summary": "Create Project",
                "description": "Criar projeto (admin)",
                "operationId": "create_project_api_portal_admin_projects_post",
                "parameters": [
                    {
                        "name": "template",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Template"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__models__client_portal__ProjectCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/app__models__client_portal__Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/portal/admin/projects/{project_id}": {
            "put": {
                "tags": [
                    "Portal do Cliente",
                    "Portal do Cliente"
                ],
                "summary": "Update Project",
                "description": "Atualizar projeto (admin)",
                "operationId": "update_project_api_portal_admin_projects__project_id__put",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Project Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__models__client_portal__ProjectUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/app__models__client_portal__Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "project": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/app__models__project__Project"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "success"
                ],
                "title": "ProjectResponse",
                "description": "Response padr\u00c3\u00a3o de Project"
            },
...
            "app__models__project__ProjectCreate": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "descricao": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Descricao"
                    },
                    "cliente_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cliente Id"
                    },
                    "cliente_nome": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cliente Nome"
                    },
                    "prioridade": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ProjectPriority"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "default": "media"
                    },
                    "prazo": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prazo"
                    },
                    "valor": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Valor"
                    },
                    "responsavel": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Responsavel"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags",
                        "default": []
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas"
                    }
                },
                "type": "object",
                "required": [
                    "nome"
                ],
                "title": "ProjectCreate",
                "description": "Payload para criar um projeto"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}